refactoring split rclone backend into stdio generic and rclone-specific
