+++ internal/backend/rest_stdio_http2/config.go
@@ -1,0 +1,5 @@
+package rest_stdio_http2
+
+type Config struct {
+	Args []string
+}
